requirements.txt
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=10.0.0
datasets>=2.14.0
transformers>=4.30.0
pytorch-forecasting>=1.0.0
pytorch-lightning>=2.0.0
autots>=0.6.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
tqdm